Removed WALG internal code dependancy on S3 interfaces
