+++ test/s3_folder_test.go
@@ -13,9 +13,9 @@ func TestS3Folder(t *testing.T) {
 	os.Setenv("AWS_ENDPOINT", "http://s3.mdst.yandex.net/")
 	os.Setenv("AWS_ACCESS_KEY_ID", "")
 	os.Setenv("AWS_SECRET_ACCESS_KEY", "")
-	os.Setenv("WALE_S3_PREFIX", "s3://test-bucket/wal-g-test-folder/Sub0")
 
-	storageFolder, err := internal.ConfigureS3Folder()
+	waleS3Prefix := "s3://test-bucket/wal-g-test-folder/Sub0"
+	storageFolder, err := internal.ConfigureS3Folder(waleS3Prefix)
 
 	assert.NoError(t, err)
